+++ compiler-core/src/type_/error.rs
@@ -15,6 +15,13 @@ use pretty_assertions::assert_eq;
 use super::FieldAccessUsage;
 
 #[derive(Debug, Eq, PartialEq, Clone)]
+pub struct UnknownType {
+    pub location: SrcSpan,
+    pub name: EcoString,
+    pub hint: UnknownTypeHint,
+}
+
+#[derive(Debug, Eq, PartialEq, Clone)]
 pub enum Error {
     SrcImportingTest {
         location: SrcSpan,
@@ -42,7 +49,7 @@ pub enum Error {
     UnknownType {
         location: SrcSpan,
         name: EcoString,
-        types: Vec<EcoString>,
+        hint: UnknownTypeHint,
     },
 
     UnknownModule {
@@ -518,11 +525,17 @@ pub fn convert_get_value_constructor_error(
     }
 }
 
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub enum UnknownTypeHint {
+    AlternativeTypes(Vec<EcoString>),
+    ValueInScopeWithSameName,
+}
+
 #[derive(Debug, PartialEq, Eq)]
 pub enum UnknownTypeConstructorError {
     Type {
         name: EcoString,
-        type_constructors: Vec<EcoString>,
+        hint: UnknownTypeHint,
     },
 
     Module {
@@ -542,13 +555,10 @@ pub fn convert_get_type_constructor_error(
     location: &SrcSpan,
 ) -> Error {
     match e {
-        UnknownTypeConstructorError::Type {
-            name,
-            type_constructors,
-        } => Error::UnknownType {
+        UnknownTypeConstructorError::Type { name, hint } => Error::UnknownType {
             location: *location,
             name,
-            types: type_constructors,
+            hint,
         },
 
         UnknownTypeConstructorError::Module {